Move legacy bookmark deletion use case to feature application module
