+++ web/src/lib/features/bookmarks/application/delete-legacy-bookmarks.ts
@@ -1,9 +1,9 @@
 import { get } from 'svelte/store';
 import { legacyBookmarkIdentifier } from '$lib/Constants';
 import { WebStorage } from '$lib/WebStorage';
-import { legacyBookmarkEvent } from './Bookmark.svelte';
-import { isLegacyBookmarkEvent } from '../features/bookmarks/domain/bookmark-migration';
-import { requestEventDeletion } from '../features/event-deletion/application/request-event-deletion';
+import { legacyBookmarkEvent } from '$lib/author/Bookmark.svelte';
+import { isLegacyBookmarkEvent } from '$lib/features/bookmarks/domain/bookmark-migration';
+import { requestEventDeletion } from '$lib/features/event-deletion/application/request-event-deletion';
 
 export async function deleteLegacyBookmarks(): Promise<void> {
 	const legacyEvent = get(legacyBookmarkEvent);
